Discards Sessions with no Signal data when appropriate
When a day has some Sessions that contain Signal data, then any Sessions that do not contain Signal data will be discarded. This was done because I had a day where most Sessions contained data, but I took a nap without the SD card in the machine and only that one Session did not have Signal data, which resulted in a very large empty space at the beginning of the Signal graphs with no explanation (this is a distinct difference from when no Sessions have Signal data, such as when the SD card was not in the machine for the entire day, in which case all Sessions are retained).